backend/agents/video_generator.py
"""
Video Generator using Veo
Creates maintenance timeline videos
"""

from google.cloud import aiplatform
from vertexai.preview.vision_models import VideoGenerationModel

class VideoGenerator:
    """Generate maintenance videos with Veo"""
    
    def __init__(self):
        self.veo_model = VideoGenerationModel.from_pretrained("veo-001")
    
    async def generate_maintenance_video(
        self,
        anomalies: list,
        predictions: dict
    ) -> str:
        """
        Generate maintenance timeline video
        
        Steps:
        1. Create script from anomalies and predictions
        2. Generate video with Veo
        3. Store in Cloud Storage
        4. Return URL
        """
        prompt = self._create_video_prompt(anomalies, predictions)
        
        video_response = self.veo_model.generate_video(
            prompt=prompt,
            duration_seconds=30
        )
        
        # Save video
        video_url = await self.storage.store_video(
            "maintenance_timeline.mp4",
            video_response.video_bytes
        )
        
        return video_url
    
    def _create_video_prompt(self, anomalies, predictions):
        """Create Veo prompt from data"""
        return f"""
        Create a 30-second industrial maintenance timeline video showing:
        - Detected anomalies at specific timestamps
        - Equipment degradation visualization
        - Recommended maintenance interventions
        - Cost savings from proactive maintenance
        
        Data: {len(anomalies)} anomalies detected, RUL: {predictions.get('rul_hours')}h
        """