requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud dependencies
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.0

# Vertex AI and Generative AI
google-generativeai==0.3.1
vertexai==1.38.0

# AI/ML dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4

# PyTorch with CUDA support (for L4 GPU)
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Transformers for Gemma
transformers==4.35.2
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# API and WebSocket
websockets==12.0
aiohttp==3.9.1
python-multipart==0.0.6

# Data processing
pyarrow==14.0.1
jsonschema==4.20.0
openpyxl==3.1.2

# Monitoring and logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1